Move Job Recovery Menu to its own file
