+++ Marlin/src/feature/power_loss_recovery.h
@@ -19,13 +19,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
+#pragma once
 
 /**
  * power_loss_recovery.h - Resume an SD print after power-loss
  */
-
-#ifndef _POWER_LOSS_RECOVERY_H_
-#define _POWER_LOSS_RECOVERY_H_
 
 #include "../sd/cardreader.h"
 #include "../core/millis_t.h"
@@ -99,5 +97,3 @@ extern uint8_t job_recovery_commands_count;
 
 void check_print_job_recovery();
 void save_job_recovery_info();
-
-#endif // _POWER_LOSS_RECOVERY_H_
